Task completed

+++ controllers/userController.js
@@ -4,15 +4,17 @@ import {
     getUserByIdQuery,
     createUserQuery,
     updateUserQuery,
-    deleteUserQuery
+    deleteUserQuery,
+    checkEmailQuery,
+    checkUserQuery,
 } from '../db/query.js';
 
 export const readAllUsers = async (req, res) => {
     try {
         const [result] = await pool.query(getAllQuery());
-        res.send(result);
+        return res.send(result);
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.status(500).json({ error: error.message });
     }
 };
 
@@ -21,19 +23,23 @@ export const readUserById = async (req, res) => {
     const { id } = req.params;
     try {
         const [result] = await pool.query(getUserByIdQuery(), [id]);
-        res.send(result);
+        return res.send(result);
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.status(500).json({ error: error.message });
     }
 };
 
 // localhost:3000/api/users
 export const createUser = async (req, res) => {
     const { name, email, address } = req.body;
-  
+
     try {
+        const checkEmail = await pool.query(checkEmailQuery(), [email]);
+        if (checkEmail[0].length > 0) {
+            return res.status(400).json({ error: 'Email already exists' });
+        }
         const [result] = await pool.query(createUserQuery(), [name, email, address]);
-        res.send(result);
+        return res.send(result);
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -41,21 +47,31 @@ export const createUser = async (req, res) => {
 
 export const updateUser = async (req, res) => {
     const { id } = req.params;
+    const existUser = await pool.query(checkUserQuery(), [id]);
+    if (existUser[0].length === 0) {
+        return res.status(404).json({ error: 'User not found' });
+    }
     const { name, email, address } = req.body;
+
     try {
         const [result] = await pool.query(updateUserQuery(), [name, email, address, id]);
-        res.send(result);
+        return res.send(result);
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.status(500).json({ error: error.message });
     }
 };
 
 export const deleteUser = async (req, res) => {
     const { id } = req.params;
+    const existUser = await pool.query(checkUserQuery(), [id]);
+    if (existUser[0].length === 0) {
+        return res.status(404).json({ error: 'User not found' });
+    }
+
     try {
         const [result] = await pool.query(deleteUserQuery(), [id]);
-        res.send(result);
+        return res.send(result);
     } catch (error) {
-        res.status(500).json({ error: error.message });
+        return res.status(500).json({ error: error.message });
     }
 };
